Moved check logic into JavaModule to also catch issues in MavenModule
